Fix code review improvements (#1)
Refactored the codebase based on a thorough code review, focusing on idiomatic Go patterns, testability, and dead code removal. All packages now have 100% test coverage.

+++ domain/operation_state.go
@@ -2,7 +2,7 @@ package domain
 
 type OperationState struct {
 	Tax                   TaxPay
-	OperationsError       int8
+	OperationsError       int
 	SharesHeld            int32
 	SharesWeightedAverage float64
 	TotalLoss             float64
